refactor: extract ql_optimizer MCP tools to separate module, switch to stdio transport
